// src/models/posfile.ts
import { ItemModel } from "./items";

export interface PosfileModel {
  [key: string]: any;
  recid?: string;
  docnum?: string;
  ordercde?: string;
  trndte?: string;
  itmcde?: string;
  itmdsc?: string;
  itmqty?: number;
  itmqty2?: number;
  trncde?: string;
  untmea?: string;
  factor?: number;
  linenum?: number;
  cuscde?: string;
  groprc?: number;
  grossprc?: number;
  usrnam?: string;
  logtim?: string;
  logdte?: Date;
  netvatprc?: number;
  exemptvat?: string;
  disper?: number;
  disamt?: number;
  amtdis?: number;
  untprc?: number;
  groext?: number;
  extprc?: number;
  netvatamt?: number;
  vatamt?: number;
  vatrte?: number;
  postrntyp?: string;
  cashier?: string;
  discde?: string;
  numpax?: number;
  distyp?: string;
  taxcde?: string;
  billdocnum?: string;
  ordocnum?: string;
  ordertyp?: string;
  cardclass?: string;
  cardtype?: string;
  cardno?: string;
  expdate?: Date;
  cardholder?: string;
  chkbnk?: string;
  chknum?: string;
  chkdte?: Date;
  freereason?: string | null;
  bankname?: string;
  approvalcode?: string;
  scharge?: number;
  scharge_disc?: number;
  vatexempt?: number;
  void?: number;
  voidnum?: string;
  customername?: string;
  address?: string;
  tin?: string;
  batchnum?: string;
  voidreason?: string;
  orderitmid?: string;
  upload_status?: string;
  postrmno?: string;
  brhcde?: string;
  memc?: string;
  memc_value?: number;
  trnsfrdte?: Date;
  trnsfrtime?: string;
  postypcde?: string;
  warcde?: string;
  isaddon?: number;
  mainitmcde?: string;
  mainitmid?: string;
  itmcomcde?: string;
  itmcomtyp?: string;
  chkcombo?: number;
  comboid?: string;
  voidqty?: number;
  itemfile?: ItemModel;
  changed?: number;
  vatamtloc?: number;
  itmpaxcount?: number;
  lessvat?: number;
  peritem?: string;
  trnstat?: number;
  nolessvat?: number;
  govdisc?: number;
  refund?: number;
  refnum?: string;
  refundreason?: string;
  refundqty?: number;
  forinv?: number;
  refunddte?: string;
  refundlogtim?: string;
  itmnum?: string;
  bnkcde?: string;
  discount?: any;
  contactno?: string;
  orderitemmodifierfiles?: any[];
  posDiscount?: {itmcde: string, cardholder: string, cardno: string, tin: string}[]
}